extracted detector base
